Crate ark_models_ext

Re-exports

pub use models::*;

Modules

fixed_base

glv

models

scalar_mul

twisted_edwards

variable_base

wnaf

Structs

Affine

Affine coordinates for a point on a twisted Edwards curve, over the base field P::BaseField.

MontgomeryAffine

Projective

Projective implements Extended Twisted Edwards Coordinates as described in [HKCD08].

Enums

TEFlags

Flags to be encoded into the serialization. The default flags (empty) should not change the binary representation.

Traits

AffineRepr

The canonical representation of an elliptic curve group element. This should represent the affine coordinates of the point corresponding to this group element.

CurveGroup

An opaque representation of an elliptic curve group element that is suitable for efficient group arithmetic.

Group

Represents (elements of) a group of prime order r.

MontCurveConfig

Constants and convenience functions that collectively define the Montgomery model of the curve. In this model, the curve equation is b * y² = x³ + a * x² + x, for constants a and b.

ScalarMul

TECurveConfig

Constants and convenience functions that collectively define the Twisted Edwards model of the curve. In this model, the curve equation is a * x² + y² = 1 + d * x² * y², for constants a and d.

VariableBaseMSM